chore(ray): add task io helper for ray
