+++ instill/configuration/__init__.py
@@ -6,9 +6,6 @@ from pathlib import Path
 import yaml
 from pydantic import BaseModel
 
-CLOUD_RAY_ADDRESS = "ray://core_ray_server:10001"
-CORE_RAY_ADDRESS = "ray://ray_server:10001"
-
 CONFIG_DIR = Path(
     os.getenv(
         "INSTILL_SYSTEM_CONFIG_PATH",
